+++ data/processed/368_wo_doc.py
@@ -71,10 +71,9 @@ class TestCases(unittest.TestCase):
                     file.write(f"This is content for {file_name}")
         for dir_name in self.dest_dirs.keys():
             os.makedirs(dir_name, exist_ok=True)
-        return super().setUp()
     def tearDown(self):
-        shutil.rmtree(self.base_test_dir)
-        return super().tearDown()
+        if os.path.exists(self.base_test_dir):
+            shutil.rmtree(self.base_test_dir)
     def test_case_1(self):
         moved_file = task_func(
             f'{self.base_test_dir}/src_test_dir_1',
